wadbuild.hs: parsec experiment
An in-progress implementation of a Parsec parser for a new wadinfo-style
format that could be used for reproducibly building/exploding WADs. E.g.

    ∀ WAD → implode(explode(WAD)) == WAD

The format would look something like

        # this is a comment
        PWAD # or IWAD
        lump rawname relative/path/to/lump
        junk external relative/path/to/lump
        junk inline "asdasd"
        wad relative/path/to/wad # sub-wad: e.g. MAP01-lumps
        pad x # shorter than "junk inline 0000".. pointless?
        label F_START # implicitly 0 length.. need an optional for a junk length argument?
        ...
        label F_END
        directory

Noteably, we will record the value of junk between lumps and in places
like the end of truncated lump names and similar.