second tool: wadxtract.hs
wadxtract <wad> <dir> explodes a WAD file <wad> into a directory <dir>.

It's very rudimentary: there are lots of problems, not least it makes no
attempt to avoid overwriting files corresponding to earlier lumps with
the same name.

Most of the interesting work is going into defining a new "wadinfo.txt" file
format, and the corresponding parser (for a wadbuild tool); but it is not
finished. The format of the wadinfo.txt file that wadxtract generates is not
representative of the finished format.

Move a pile of common code out of lswad.hs into wad.hs.