wadcat: change output method to stdout
All command-line supplied filenames are treated as inputs. Write
the combined PWAD to standard out. These semantics match the UNIX
"cat" program that inspires the name.

Additionally, check to ensure that stdout has been redirected, so
we don't write the PWAD to an interactive terminal. Fail the program
with an explanation if the user has not redirected the output.